# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Layanan Pajak API",
        "description": "REST API untuk data realisasi pajak daerah Kabupaten Pasuruan. Semua endpoint memerlukan Bearer Token pada header Authorization.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://upp.pendapatan.pasuruankab.go.id/api",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development"
        }
    ],
    "paths": {
        "/v1/payment-check/hotel": {
            "post": {
                "tags": [
                    "Cek Pembayaran"
                ],
                "summary": "Cek pembayaran PBJT Hotel",
                "operationId": "2249fac82b64f17948d7f7e7a3ee9195",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentCheckBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentCheckResult"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/v1/payment-check/restoran": {
            "post": {
                "tags": [
                    "Cek Pembayaran"
                ],
                "summary": "Cek pembayaran PBJT Restoran",
                "operationId": "a737539bfae992fc00038732ad33cdb6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentCheckBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentCheckResult"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/v1/payment-check/hiburan": {
            "post": {
                "tags": [
                    "Cek Pembayaran"
                ],
                "summary": "Cek pembayaran PBJT Hiburan",
                "operationId": "d8b08bea93f8e3dcf94508cc6c2b22da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentCheckBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentCheckResult"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/v1/payment-check/reklame": {
            "post": {
                "tags": [
                    "Cek Pembayaran"
                ],
                "summary": "Cek pembayaran Pajak Reklame",
                "operationId": "6fe6077b0884bdc4f849dc18cd4c33a0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentCheckBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentCheckResult"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/v1/payment-check/ppj": {
            "post": {
                "tags": [
                    "Cek Pembayaran"
                ],
                "summary": "Cek pembayaran Pajak Penerangan Jalan (PPJ)",
                "operationId": "ed1851ae5fe30148d8d225aeeee84ffb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentCheckBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentCheckResult"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/v1/payment-check/parkir": {
            "post": {
                "tags": [
                    "Cek Pembayaran"
                ],
                "summary": "Cek pembayaran PBJT Parkir",
                "operationId": "9cc7e1d5ca591ce6ffd22cc2d84cd990",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentCheckBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentCheckResult"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/v1/payment-check/at": {
            "post": {
                "tags": [
                    "Cek Pembayaran"
                ],
                "summary": "Cek pembayaran Pajak Air Tanah (AT)",
                "operationId": "8c3a7894ba70827cf3dd82a4d0f29fb6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentCheckBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentCheckResult"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/v1/payment-check/minerba": {
            "post": {
                "tags": [
                    "Cek Pembayaran"
                ],
                "summary": "Cek pembayaran Pajak Mineral Bukan Logam (MINERBA)",
                "operationId": "cd90103e123cc09e0c55bbddda0e13a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentCheckBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentCheckResult"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/v1/payment-check/bphtb": {
            "post": {
                "tags": [
                    "Cek Pembayaran"
                ],
                "summary": "Cek pembayaran BPHTB",
                "description": "Untuk BPHTB, field `npwpd` diisi dengan Nomor Register (kohir).",
                "operationId": "007a7c83b40685acc2b8cf206c406d0e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentCheckBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentCheckResult"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaymentCheckItem": {
                "properties": {
                    "nop": {
                        "type": "string"
                    },
                    "nama_op": {
                        "type": "string"
                    },
                    "kohir": {
                        "type": "string"
                    },
                    "masa_awal": {
                        "type": "string"
                    },
                    "jatuh_tempo": {
                        "type": "string",
                        "nullable": true
                    },
                    "tgl_bayar": {
                        "type": "string",
                        "nullable": true
                    },
                    "jml_sptpd": {
                        "type": "integer"
                    },
                    "bayar_pokok": {
                        "type": "integer"
                    },
                    "bayar_denda": {
                        "type": "integer"
                    },
                    "sisa_pokok": {
                        "type": "integer"
                    },
                    "sisa_denda": {
                        "type": "integer"
                    },
                    "sisa": {
                        "type": "integer"
                    },
                    "keterangan": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaymentCheckResult": {
                "properties": {
                    "npwpd": {
                        "type": "string",
                        "example": "1234567890123"
                    },
                    "nama_op": {
                        "type": "string",
                        "example": "HOTEL SEJAHTERA"
                    },
                    "jenis_pajak": {
                        "type": "string",
                        "example": "hotel"
                    },
                    "tahun": {
                        "type": "string",
                        "example": "2025"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PaymentCheckItem"
                        }
                    },
                    "summary": {
                        "properties": {
                            "total_sptpd": {
                                "type": "integer"
                            },
                            "total_denda": {
                                "type": "integer"
                            },
                            "total_bayar_pokok": {
                                "type": "integer"
                            },
                            "total_bayar_denda": {
                                "type": "integer"
                            },
                            "total_sisa": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PaymentCheckBody": {
                "required": [
                    "npwpd",
                    "tahun"
                ],
                "properties": {
                    "npwpd": {
                        "description": "Nomor Pokok Wajib Pajak Daerah",
                        "type": "string",
                        "example": "1234567890123"
                    },
                    "tahun": {
                        "description": "Tahun pajak",
                        "type": "integer",
                        "example": 2025
                    },
                    "npwpd_lama": {
                        "description": "true jika menggunakan NPWPD lama",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "description": "Masukkan token API.",
                "scheme": "bearer"
            }
        }
    },
...
}